Extract Kore.Equation.Registry
