+++ kore/src/Kore/Equation.hs
@@ -7,7 +7,9 @@ License     : NCSA
 module Kore.Equation
     ( module Kore.Equation.Equation
     , fromSentenceAxiom
+    , extractEquations
     ) where
 
 import Kore.Equation.Equation
+import Kore.Equation.Registry
 import Kore.Equation.Sentence
